removed all admin related components and pages
