+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@clerk/nextjs": "^4.29.9",
+    "@prisma/client": "^5.13.0",
     "@radix-ui/react-icons": "^1.3.0",
     "@vercel/analytics": "^1.2.2",
     "draft-js": "^0.11.7",
@@ -24,6 +25,7 @@
     "sonner": "^1.4.3"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.13",
     "@types/draft-js": "^0.11.17",
     "@types/node": "^20",
     "@types/react": "^18",
