+++ prisma/schema.prisma
@@ -1,0 +1,55 @@
+datasource db {
+  provider = "mongodb"
+  url      = env("DATABASE_URL")
+}
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+model User {
+  id      String   @id @default(auto()) @map("_id") @db.ObjectId
+  email   String   @unique
+  name    String?
+  address Address?
+  articles Article[]
+  events Event[]
+}
+
+model Category {
+  id String @id @default(auto()) @map("_id") @db.ObjectId
+  title String
+  articles Article[]
+  slug String @unique
+}
+
+model Article {
+  id       String    @id @default(auto()) @map("_id") @db.ObjectId
+  createdAt DateTime @default(now())
+  slug     String    @unique
+  title    String
+  excerpt String
+  body     String
+  author   User      @relation(fields: [authorId], references: [id])
+  authorId String    @db.ObjectId
+  catSlug String 
+  category Category @relation(fields: [catSlug], references: [slug])
+}
+
+model Event {
+  id       String    @id @default(auto()) @map("_id") @db.ObjectId
+  slug     String    @unique
+  title    String
+  description     String
+  date String
+  author   User      @relation(fields: [authorId], references: [id])
+  authorId String    @db.ObjectId
+}
+
+// Address is an embedded document
+type Address {
+  street String
+  city   String
+  state  String
+  zip    String
+}
