+++ src/app/(auth)/[[...sign-up]]/legacy.page.tsx
