+++ src/app/(auth)/[[...sign-up]]/page.tsx
