--- src/app/admin/authors/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Authors() {
-  return (
-    <div className='min-h-[92vh] grid place-items-center bg-gray-50'>
-      <h1 className='font-bold text-2xl lg:text-4xl lg:mb-5'>Authors</h1>
-    </div>
-  );
-}
-
-export default Authors;
