--- src/app/admin/create/headline/page.tsx
@@ -1,9 +1,0 @@
-function Page() {
-  return (
-    <section className='grid h-full place-content-center'>
-      <h1>Headline</h1>
-    </section>
-  );
-}
-
-export default Page;
