--- src/app/admin/create/opportunity/page.tsx
@@ -1,9 +1,0 @@
-function Page() {
-  return (
-    <section className='grid h-full place-content-center'>
-      <h1>Opportunity</h1>
-    </section>
-  );
-}
-
-export default Page;
