--- src/app/admin/help/page.tsx
@@ -1,9 +1,0 @@
-function Page() {
-  return (
-    <div className='grid place-content-center min-h-[calc(100vh-60px)]'>
-      <h1>Tutorial for using the admin app</h1>
-    </div>
-  );
-}
-
-export default Page;
