--- src/app/providers.tsx
@@ -1,9 +1,0 @@
-import Footer from '@/components/shared/main-footer';
-import Navbar from '@/components/shared/navigation';
-import AuthProvider from '@/context/AuthProvider';
-
-function Providers({ children }: { children: React.ReactNode }) {
-  return <></>;
-}
-
-export default Providers;
