--- src/context/AuthContext.tsx
@@ -1,17 +1,0 @@
-import { AuthContextValues } from '@/types';
-import { createContext } from 'react';
-
-export const INITIAL_STATE = {
-  authenticated: false,
-  user: null,
-};
-
-const AuthContext = createContext<AuthContextValues>({
-  user: null,
-  token: null,
-  authenticated: false,
-  logout : () => {},
-  authenticate: () => {},
-});
-
-export default AuthContext;
