--- src/lib/mongoose.ts
@@ -1,25 +1,0 @@
-// import mongoose from "mongoose";
-
-// let isConnected: boolean = false;
-
-// export const connectToDb = async () => {
-//   mongoose.set("strictQuery", true);
-
-//   if (!process.env.MONGO_DB_URI) {
-//     return console.log("Missing MONGO URI");
-//   }
-
-//   if (isConnected) {
-//     return console.log("MongoDB is already connected");
-//   }
-
-//   try {
-//     await mongoose.connect(process.env.MONGO_DB_URI, { dbName: "devflow" });
-
-//     isConnected = true;
-
-//     console.log("MongoDB connected");
-//   } catch (error) {
-//     console.log("MongoDB connection error: ", error);
-//   }
-// };
