+++ tailwind.config.ts
@@ -46,6 +46,6 @@ const config: Config = {
       },
     },
   },
-  plugins: [],
+  plugins: [ require('@tailwindcss/typography'),],
 };
 export default config;
